2 methods in Owner class: get_owners_accounts_ids and return_owners_accounts
To create the relationship between the accounts and the owners use an account_owners.csv file. The data for this file, in order in the CSV, consists of: 1. Account ID - (Fixnum) a unique identifier corresponding to an Account instance. 1. Owner ID - (Fixnum) a unique identifier corresponding to an Owner instance.

get_owners_accounts_ids uses the CSV file to return an array of all account ids associated with the owner instance's owner id.
return_owners_accounts translates these account ids into an array of account instances (with the matching account ids) using the class method Account.find(id)