tried moving interest rate to account.
since the code is used in both Savings and MoneyMarket, it makes sense to move it to the parent class. I considered using include instead, but I think this is a behavior common to enough account types that it is an Account behavior.

(commented it out in MM and Savings until I've tested.)